Make models use `Decimal` for prices instead of `float`
This does not fix an existing bug, because no arithmetic was being
performed for the fields in question, but for future safety reasons this
is the correct type to use.

+++ models/transactions.py
@@ -1,5 +1,6 @@
 import datetime
 from dataclasses import dataclass
+from decimal import Decimal
 
 
 @dataclass
@@ -20,6 +21,5 @@ class Customer:
 class Transaction:
     bike_id: int
     customer_id: int
-    # TODO: maybe Decimal?
-    total_cost: float
+    total_cost: Decimal
     date: datetime.date
